+++ tools/demo/warehouse_publisher.py
@@ -1,11 +1,11 @@
 import json
 import os
 import argparse
+from datetime import datetime, UTC
+from uuid import uuid4
 
 from lab_share_lib.config.rabbit_server_details import RabbitServerDetails
 from lab_share_lib.rabbit.basic_publisher import BasicPublisher
-
-from tools.demo.create_aliquot_in_mlwh_messages import build_create_aliquot_message
 
 RABBITMQ_HOST = os.getenv("RABBITMQ_HOST", "localhost")
 RABBITMQ_PORT = os.getenv("RABBITMQ_PORT", "5671")
@@ -15,11 +15,35 @@ RABBITMQ_VHOST = os.getenv("RABBITMQ_VHOST", "test")
 RABBITMQ_EXCHANGE = os.getenv("RABBITMQ_EXCHANGE", "psd.tol-lab-share")
 RABBITMQ_ROUTING_KEY = os.getenv("RABBITMQ_ROUTING_KEY", "development.saved.aliquots.*")
 
-if __name__ == '__main__':
+
+def build_create_aliquot_message():
+    dt = datetime.now(UTC)
+    date_string = dt.strftime("%Y-%m-%dT%H:%M:%SZ")
+
+    return {
+        "lims": "traction",
+        "aliquot": {
+            "id_lims": "LIMS123456",
+            "lims_uuid": str(uuid4()),
+            "aliquot_type": "derived",
+            "source_type": "primary",
+            "source_barcode": "SRC123456",
+            "sample_name": "SampleA",
+            "used_by_type": "Research",
+            "used_by_barcode": "USR123456",
+            "volume": 50.50,
+            "concentration": 200.10,
+            "last_updated": date_string,
+            "recorded_at": date_string,
+            "created_at": date_string,
+            "insert_size": 350,
+        },
+    }
+
+
+if __name__ == "__main__":
     parser = argparse.ArgumentParser(description="TOL LabShare message publisher demo script.")
-    parser.add_argument(
-        "--routing_key", default="development.saved.aliquots.1", help="Routing key for the binding"
-    )
+    parser.add_argument("--routing_key", default="development.saved.aliquots.1", help="Routing key for the binding")
 
     args = parser.parse_args()
 
@@ -29,15 +53,10 @@ if __name__ == '__main__':
         port=RABBITMQ_PORT,
         username=RABBITMQ_USERNAME,
         password=RABBITMQ_PASSWORD,
-        vhost=RABBITMQ_VHOST
+        vhost=RABBITMQ_VHOST,
     )
 
-    publisher = BasicPublisher(
-        rmq_details,
-        publish_retry_delay=5,
-        publish_max_retries=36,
-        verify_cert=False
-    )
+    publisher = BasicPublisher(rmq_details, publish_retry_delay=5, publish_max_retries=36, verify_cert=False)
 
     message = build_create_aliquot_message()
     publisher.publish_message(
@@ -48,9 +67,3 @@ if __name__ == '__main__':
         schema_version=None,
         encoder_type=None,
     )
-    
-
-
-
-
-
